+++ gap/apps/pricelist/models.py
@@ -29,11 +29,9 @@ class Price(models.Model):
         max_digits=10, decimal_places=2, validators=[MinValueValidator(0)],
         verbose_name='RPL Price')
 
-    quantity = models.DecimalField(max_digits=10, decimal_places=2,
-                                   validators=[MinValueValidator(0)])
+    quantity = models.IntegerField(validators=[MinValueValidator(0)])
 
-    min_order = models.DecimalField(max_digits=10, decimal_places=2,
-                                    validators=[MinValueValidator(0)])
+    min_order = models.IntegerField(validators=[MinValueValidator(0)])
 
     option_choices = models.ManyToManyField(
         OptionChoice, related_name='prices', blank=True,
